refactor: move publish tabs
