--- src/app/dashboard/publish/layout.tsx
@@ -1,35 +1,0 @@
-"use client";
-
-import { useMemo } from "react";
-import {
-  UrqlProvider,
-  ssrExchange,
-  cacheExchange,
-  fetchExchange,
-  createClient,
-} from "@urql/next";
-
-export default function DashboardLayout({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
-  const [client, ssr] = useMemo(() => {
-    const ssr = ssrExchange({
-      isClient: typeof window !== "undefined",
-    });
-    const client = createClient({
-      url: process.env.NEXT_PUBLIC_GRAPHQL_URL!,
-      exchanges: [cacheExchange, ssr, fetchExchange],
-      suspense: true,
-    });
-
-    return [client, ssr];
-  }, []);
-
-  return (
-    <UrqlProvider client={client} ssr={ssr}>
-      {children}
-    </UrqlProvider>
-  );
-}
